requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
starlette==0.47.2
uvicorn==0.35.0
openai>=1.40.0
python-dotenv